+++ actionpack/lib/action_dispatch/middleware/executor.rb
@@ -13,9 +13,6 @@ module ActionDispatch
       begin
         response = @app.call(env)
         returned = response << ::Rack::BodyProxy.new(response.pop) { state.complete! }
-      rescue => error
-        @executor.error_reporter.report(error, handled: false, source: "application.action_dispatch")
-        raise
       ensure
         state.complete! unless returned
       end